Update app.js added updates to claude see notes
I apologize for any confusion. Let me clarify the status of these functions:

async function updateGameModeAndMapOptions():
This is a new function that needs to be added to your app.js file. It's not present in your current code.
function updatePlacementInput():
This function already exists in your current code, but it needs to be updated. The version I provided is an updated version that accounts for the new game structure.
function updatePlacementValue():
This function already exists in your current code and doesn't need to be changed. I included it for completeness, but you can keep your existing version.

To implement these changes:

Add the new updateGameModeAndMapOptions() function to your app.js file.
Replace your existing updatePlacementInput() function with the new version I provided.
Keep your existing updatePlacementValue() function as it is.

I apologize for not being clearer about which functions were new and which were updates. In the future, I'll make sure to explicitly state whether a function is new or an update to an existing one. Would you like me to provide any further clarification or assistance with implementing these changes?